+++ ogn/parser/aprs_comment/generic_parser.py
@@ -2,8 +2,8 @@ from .base import BaseParser
 
 
 class GenericParser(BaseParser):
-    def __init__(self):
-        self.beacon_type = 'generic'
+    def __init__(self, beacon_type='unknown'):
+        self.beacon_type = beacon_type
 
     def parse_position(self, aprs_comment):
         return {'comment': aprs_comment}